Limit repository search to alphagov org only
Searching repositories is a fairly expensive operation and currently
only scoped to "topic:govuk". This will return repositories that have a
govuk topic, but sit outside of the alphagov organisation

We don't need to worry about repositories that sit outside of our
organisation for the purposes of surfacing our dependencies.

Specifying the org name when getting a list of repositories will leave
us with a smaller set to enumerate over during the rest of our code. It
also has a nice side effect that speeds up the search a little.

Without filtering on org:

```
puts Benchmark.measure {
  10.times do
    octokit.search_repos("topic:govuk")
  end
}

```

With filtering on org:

```
puts Benchmark.measure {
  10.times do
    octokit.search_repos("org:alphagov topic:govuk")
  end
}

```