Create task to re-record cassettes
VCR cassettes should be re-recorded when searches against the api
change. This adds a rake task to re-record the cassettes.

In theory simply deleting the cassette and re-running the tests should
re-record the cassette. It does do this, but I kept hitting an issue
where the cassette would no longer represent a 200 against the github
api and instead returned a 401 unathorized.

This is possibly because of the inclusion of a fake token in
spec/gateways/repositories_spec.rb where we pass
`GITHUB_TOKEN: "some_token"` in the `ClimateControl` block.

I considered complicating the tests to selectively pass the
token when the repository cassette did not exist, but that felt like
complexity in a test that didn't belong there. Instead I've opted to
create a rake task to re-record the cassette and accepted that there
will be some duplication of the github query (once in the rakefile and
once in the repository gateway). I decided not to create a constant for
this because I feel the code is complicated enough already.