Fix linting errors
Notes on the peculiarities:

Using `open` is considered a vulnerability because Kernel.open can open
processes as well as files. So we should use File.open in
`lib/gateways/file.rb`. Unfortunately that tries to find `Gatways::File#open`
and so, we end up having to use `::File` to get to the top level Ruby
namespace.