Remove fragile dependency on a file in security test
The security test relies on running bundler-audit against a file that
should represent a Gemfile without any security vulnerabilities.

Using a fixture to represent this physical file is certainly one way to
do this, but suffers from the disadvantage of the file becoming out of
date fairly easily as new vulnerabilities are found.

This means that we would need to regularly update the fixture as we
enhance the code base. This seems fairly tedious.

Using a fixture does let us run a real test against bundler-audit, but
lets rather define the boundary of our application as the point up to
which we test. We can stub the response from bundler audit to an empty
array when we want to test the cases of no security vulnerabilities
being present. This means we are really only testing our code and how we
react to and work with an empty array being returned from the #scan
method that bundler-audit exposes to us. I think this is fine, our
responsibility should be to test our code, not exercise the code paths
in third party code.

There is probably a way to update the fixtures fairly gracefully, but it
seemed cumbersome and the docs don't really give a good indication of
how to do so automatically, so I think it would involve actually
rummaging through and finding a file that has no vulnerabilities and
commiting that to the repo. That seems fragile because between opening
the PR and merging to master after approval, a new vulnerability could
be discovered and would potentially render the test useless again.