Remove fragile dependency on a file in security test
Testing our security features relies on running bundler-audit
against a fixtures that should represent a Gemfile without any security
vulnerabilities.

Using a fixture to represent this physical file is certainly one way to
do this, but suffers from the disadvantage of the file becoming out of
date fairly easily as new vulnerabilities are found.

This means that we would need to regularly update the fixture as we
enhance the code base. This seems fairly tedious.

Using a fixture does let us run a real test against bundler-audit which
exercise the code through it's full journey. However, bundler-audit is a
gem and it's not really our concern. We can rather stub the response
that we are expecting from bundler-audit. This means we confine the
boundary of our tests to our code only. The disadvantage of stubbing out
the response we expect is that should bundler-audit's interface and
response change, we tests that make assumptions about the return value
which could pass, even though the actual implementation would fail. I
think this trade off is fine. We are relying on bundler-audit's public
interface and should that change it would be a breaking change in the
software and more than likely indicated in change-notes.

Although we could go through the process of updating the fixtures, I
think it represents a worse approach than subbing because there are
multiple points where it could fail. The time between pushing code and
opening a PR could represent a point in time in which a previously
unknown vulnerability is found. This would  break our tests. The same
could be said for the time between a PR being approved and merged (when
our tests run again). The failures from the fixture suddenly
representing security vulnerabilities whene there were none during the
code cycle is unexpected and just causes developer confusion.

I've lef the fixtures for the other cases for now because I don't think
they represent such a fragile test (for example fixtures that represent
Gemfiles with vulnerabilities are fine - they will not go out date).