Fix linting
The only odd thing in this commit is the use of `::File`
in `lib/gateways/file.rb`. This is because it's inside the
`Gateways::File` class so using `File.open` tries to find
`Gateways::File#open` which doesn't exits. Using `::File.open` calls the
method on the base namespace in Ruby.