Add use case to distribute overflow
To help distribute dependabot PRs a little more fairly around the
GOV.UK developer community, we want to be able to move some overflow
from teams into the general govuk-developers channel.

Add a use case to do this. The algorithm is not very clever, it looks
at the number of applications with open pull requests per team and
if there are more than a defined overflow of applications, it pops these
from the overworked team and concats them onto the govuk-developers team.

A better algorithm might be to look at the combined number of pull
requests per team and then distribute individual pull requests so a team
doesn't have to contend with more than n number of pull requests per
application. That would, however, require the slack message posting to
change up a little so that we filter a little more accurately when
someone clicks a link through to Github. At the moment, clicking a link
in the slack message will take you to all dependabot prs for the
application you've clicked, so it felt like I would be reworking a
larger portion of the application if I went down that route.