Move VCR out of the test group
When this project is built it is bundled without the test group. This
causes the rake tasks to fail because we now require vcr in the Rakefile
to help with re-creating the recordings.

We could alternatively move the re-record task out of the Rakefile, but
this should get us going again so we at least have visibility of the
dependabot PRs.

Also add /.bundle to the gitignore