Add `continuously_deployed_apps` property to `Gateways::Team`
We will add the `continuously_deployed` property to the Developer
Docs prior to merge. The `continuously_deployed_apps` will be used
to edit the Dependabot Slack messages, drawing attention to PRs
which can be merged & deployed with one click, thanks to the
Continous Deployment process. This should mean developers are more
open to merging Dependabot PRs.