Count number of PRs open in CD-enabled apps before deciding string
It's quite possible that there are CD-enabled apps managed by the
given team, but that there are no PRs open for them. Currently, it
will say:

```
[TEAM_NAME] have 0 Dependabot PRs open on the following
Continuously Deployed apps:

And 73 Dependabot PRs open on other apps:
```

Screenshot:

![](https://user-images.githubusercontent.com/5111927/107197652-10346680-69ec-11eb-926d-64b8f02f74a5.png)

This commit now counts the number of PRs in those apps before
deciding whether to show the CD part of the string.