Pass 'continuously_deployed_apps' to govuk-developers overflow
govuk-developers is an edge case whereby if we have too many PRs
for a team(s) to handle, it 'overflows' into #govuk-developers.
This meant that `:continuously_deployed_apps` wasn't being defined
for the #govuk-developers entry.

This commit ensures that any continuously-deployed-apps
information is passed through with the overflow, so that the
govuk-developers channel gets the same CD improvement as the team
specific ones.