Define rake tasks to warm and clear the cache
Heroku times out waiting for github to respond, as the requests are so spaced
out by the sleeps added in previous commits. We still need to fetch the data
from github with sleeps to prevent being rate limited. But rather than do that
in real time whilst trying to load the page we should move that to an async task.

The warm cache task will be kicked off by heroku scheduler twice a day.