Use a filestore for development and test environments
Running a local memcached server in order to load the app or run the
tests is a pain, so use a filestore unless it's production.

The api for dalli is fetch/set/get, but it's fetch/write/read for
activesupport::filestore. Likewise to clear the cache, the command
for dalli is cache.clear, but for activesupport its cache.flush_all.