Remove all Continuous Deployment related logic
The 'continuously deployed apps' logic ([1]) has been broken for a long time.
[1]: https://github.com/alphagov/govuk-dependencies/blob/e245152349ed4428ead26de42ca3e344efd454b5/lib/presenters/slack/full_message.rb#L7-L14

This is because govuk-dependencies tries to read that info from the Developer Docs ([2]).
[2]: https://github.com/alphagov/govuk-dependencies/blob/6d9dfbae47e5e866d676cb2285e0fef93406153d/lib/gateways/team.rb#L7-L16

However, we removed that from the Developer Docs in October 2021 ([3]).
[3]: https://github.com/alphagov/govuk-developer-docs/commit/5f667af9ae00abf969a80408358409eeca97e7b1

Whilst we could in theory retrieve that info from the Release app now ([4]), not
all repos retrieved by govuk-dependencies has a corresponding app in the Release
app, and the JSON endpoints we'd have to query are behind Signon.
[4]: https://release.publishing.service.gov.uk/applications/account-api.json

We did notice the change in govuk-dependencies Slack messages after the move
of data to the Release app, but the technical difficulties above, coupled
with the fact that more and more apps are being migrated to use Continuous
Deployment (and the effort being better spent there) meant we decided not to
fix it.

We kept the code around on the basis it might be useful someday, but we can
always retrieve it from version control if needed, and it currently adds
quite a bit of complexity to the govuk-dependencies app, so is better off
removed.