Commit Apps Script to GitHub
We have this Apps Script 'installed' on the following Google sheet:
https://docs.google.com/spreadsheets/d/1Lt-uarJm1xaCEC58KdSixsa5LsBSmNXym37SbuXbrcI/edit

It allows you to retrieve the latest repo list from the Developer
Docs, and then retrieve the versions of some key dependencies for
each repo (Ruby, Rails, Sidekiq, etc).

Note that the Apps Script is a JavaScript based platform:
https://developers.google.com/apps-script

This code isn't directly related to govuk-dependencies, but is
very closely linked to what govuk-dependencies aims to do. So
rather than create a new repo just for one file, we're adding it
here.