Reduce scope of permissions granted via authorisation dialog
When running the script for the first time, the user has to grant
permission to the script. Without this annotation, the permission
extends to all spreadsheets the user has access to. With this
annotation, permission is only granted to access the current
spreadsheet, which is safer.

See https://developers.google.com/apps-script/guides/services/authorization#manual_authorization_scopes_for_sheets_docs_slides_and_forms